Allow to specify a name for the repository
Normally, the name of the repository is detected via the argument
name that is given on the command line. But, in some cases, this
argument is a relative path, or something like ".". A particular
case is when running the inspector in a Docker container, where
the repository is stored in a generic directory like "repo".

This adds a simple option to configure that name. This will *not*
work properly when giving multiple repositories on the command line.
Hopefully, it should give the same name to all the repositories.
In the worst case, there will be clashes somewhere in the code.